Move check_swarm_mode from scanner to dockerclient object
